Fixing test.
No occurrences will be returned if the event starts before the
calculated query range in the recurrence module. This is true, if we
leave start and end dates in the tests to be dynamically calculated
while the test is run. Therefore I manually create events which start
later, than the calculated query range.