Add event_uid field, with index, to be used for ical sync
The ical spec requires a UID field to be set for syncing [1], the
recommended format is a UID specific to the host followed by "@" and
the host domain.

Note: this work was cancelled mid-flight because of a change in
requirements, it doesn't do much yet and is probably broken.

The idea was to set the event_uid on import. When syncing, an existing
event with the same event_uid could have it's fields updated to match
the current values in the ical feed.

1. http://www.kanzaki.com/docs/ical/uid.html