Attempt to fix table in pypi
Both locally and on github, there is no problem, but if you look at the compatibility table on pypi, the first heading row is off:  `Plone` should be shifted one column to the right.  I'm not sure if this change will fix the issue, but it doesn't affect the local or the github rendering.  Only way to know, that I can think of, is to wait for the next release on pypi.
(Also, I don't think this needs a changelog entry.)